requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.12.0
evaluate>=0.4.0
scipy>=1.10.0
pillow>=9.5.0
numpy>=1.24.0
tqdm>=4.65.0
wandb>=0.15.0
pytorch-fid>=0.3.0
scikit-learn>=1.2.0
matplotlib>=3.7.0 